# src/server/api-doc.yaml
openapi: 3.0.2
info:
  description: "List of APIs exposed by the test server"
  version: "1.0.0"
  title: "Test API Server"
paths:
...
  "/api/stocks":
    get:
      summary: Returns the quote for a given U.S. stock symbol
      operationId: getQuote
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful request
  "/api/users/validate":
    get:
      summary: Validates user input
      operationId: validateUser
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful request
...